fix: set children to null when needed & re-add parts of the template (#129)
These parts of the template were removed because generation was getting
stuck. The issue was that the template engine "looks up" for existing
variables they may not exist in the current scope.

In this case, we loop through `{{#children}}` several times. But, the
`children` variable never gets updated. So, we end up nesting the same
loop several times, which takes a really long time to render.

Removing stuff from the loop made it faster to render, but still led to
a lot of duplicate content.