fix: Run docfx from a local copy when testing
This is a horrible hack that needs to be addressed properly.
In order to run plugins, docfx creates a subdirectory under the
docfx executable's directory. That fails when docfx is in a
read-only directory. This is really a design issue in docfx, but I
don't expect that to be fixed imminently.

In order to run plugins in the docfx pipeline, we'll either need to
do something like this hack in doc-pipeline as well, or change the
Dockerfile to make /opt/docfx writable.